[chore] : Remove unnecessary files from libs dir
