--- libs/supabase.js
@@ -1,9 +1,0 @@
-import { createBrowserClient } from "@supabase/ssr";
-
-const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL;
-const supabaseAnonKey = process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY;
-
-
-export const createClient = () => {
-  return createBrowserClient(supabaseUrl, supabaseAnonKey);
-};
